changes after latest main merge with issues
