+++ pages/my-reports/style.css
@@ -1,5 +1,5 @@
 @import url(../../assets/css/report-card.css);
-@import url(../../assets/css/switch.css);
+@import url(../../assets/css/toggle-switch.css);
 
 #recentReports,
 #olderReports {
